Ice Potion and Fire Potion

+++ Codebase/1906WorkingTitle/Assets/Scripts/EnemyStats.cs
@@ -9,7 +9,7 @@ public class EnemyStats : MonoBehaviour
     [SerializeField] private int enemyPoints = 1;
 
     //How many hits it takes to kill the enemy.
-    [SerializeField] private int health = 2;
+    [SerializeField] private float health = 2;
 
     //How much damage the enemy deals on hit.
     [SerializeField] private int damage = 2;
@@ -43,7 +43,7 @@ public class EnemyStats : MonoBehaviour
     {
         return enemyPoints;
     }
-    public int GetHealth()
+    public float GetHealth()
     {
         return health;
     }
@@ -64,7 +64,7 @@ public class EnemyStats : MonoBehaviour
         return bulletSpeed;
     }
 
-    public void SetHealth(int _health)
+    public void SetHealth(float _health)
     {
         health = _health;
     }
@@ -87,7 +87,7 @@ public class EnemyStats : MonoBehaviour
     #endregion getset
 
     //Our enemy is damaged
-    public void TakeDamage(int _damage = 1)
+    public void TakeDamage(float _damage = 1)
     {
         if(damage > 0)
         {
